# api/providers/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Provider API",
    "description": "Service to manager all provider related data: delivery zones and slots, pickup points, tariffs, etc.",
    "version": "1.0"
  },
  "tags": [
    {
      "name": "ProvidersService"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/providers": {
      "get": {
        "summary": "List all providers",
        "operationId": "ProvidersService_ProviderListAll",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ProviderListAllResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ProvidersService"
        ]
      },
      "post": {
        "summary": "Create a provider",
        "operationId": "ProvidersService_ProviderCreate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ProviderCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Creates a new delivery provider",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ProviderCreateRequest"
            }
          }
        ],
        "tags": [
          "ProvidersService"
        ]
      }
    },
    "/v1/providers/{provider_id}": {
      "delete": {
        "summary": "Delete a provider",
        "operationId": "ProvidersService_ProviderDelete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ProviderDeleteResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ProvidersService"
        ]
      },
      "put": {
        "summary": "Update a provider",
        "operationId": "ProvidersService_ProviderUpdate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ProviderUpdateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "provider_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProvidersServiceProviderUpdateBody"
            }
          }
        ],
        "tags": [
          "ProvidersService"
        ]
      }
    }
  },
  "definitions": {
    "ProvidersServiceProviderUpdateBody": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Provider": {
      "type": "object",
      "properties": {
        "provider_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "v1ProviderCreateRequest": {
      "type": "object",
      "properties": {
        "provider_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      },
      "description": "Creates a new delivery provider",
      "title": "ProviderCreateRequest",
      "required": [
        "provider_id",
        "name"
      ]
    },
    "v1ProviderCreateResponse": {
      "type": "object",
      "properties": {
        "provider_id": {
          "type": "string"
        }
      },
      "title": "ProviderCreateResponse"
    },
    "v1ProviderDeleteResponse": {
      "type": "object"
    },
    "v1ProviderListAllResponse": {
      "type": "object",
      "properties": {
        "providers": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Provider"
          }
        }
      }
    },
    "v1ProviderUpdateResponse": {
      "type": "object",
      "properties": {
        "provider": {
          "$ref": "#/definitions/v1Provider"
        }
      }
    }
  }
}